+++ lib/src/constraint/flexible_table_row_height.dart
@@ -8,39 +8,39 @@ abstract class AbsFlexibleTableRowHeight<T> {
   double get headerRowHeight;
 
   ///表信息行高度
-  double? get infoRowHeight;
+  double? get fixedInfoRowHeight;
 
   double? Function(int dataIndex, T data)? get infoRowHeightBuilder;
 
   ///信息行的固定高度
-  double fixedInfoRowHeight(FlexibleTableController<T> controller, int dataIndex, T data);
+  double infoRowHeight(FlexibleTableController<T> controller, int dataIndex, T data);
 }
 
 ///固定高度
 class FixedTableRowHeight<T> extends AbsFlexibleTableRowHeight<T> {
   const FixedTableRowHeight({
     required this.headerRowHeight,
-    required this.infoRowHeight,
+    required this.fixedInfoRowHeight,
   });
 
   @override
   final double headerRowHeight;
 
   @override
-  final double infoRowHeight;
+  final double fixedInfoRowHeight;
 
   @override
   double? Function(int dataIndex, T data)? get infoRowHeightBuilder => null;
 
   @override
-  double fixedInfoRowHeight(FlexibleTableController<T> controller, int dataIndex, T data) => infoRowHeight;
+  double infoRowHeight(FlexibleTableController<T> controller, int dataIndex, T data) => fixedInfoRowHeight;
 }
 
 ///根据数据变化
 class ChangeableTableRowHeight<T> extends AbsFlexibleTableRowHeight<T> {
   const ChangeableTableRowHeight({
     required this.headerRowHeight,
-    this.infoRowHeight,
+    this.fixedInfoRowHeight,
     required this.infoRowHeightBuilder,
   });
 
@@ -48,13 +48,13 @@ class ChangeableTableRowHeight<T> extends AbsFlexibleTableRowHeight<T> {
   final double headerRowHeight;
 
   @override
-  final double? infoRowHeight;
+  final double? fixedInfoRowHeight;
 
   @override
   final double? Function(int dataIndex, T data) infoRowHeightBuilder;
 
   @override
-  double fixedInfoRowHeight(FlexibleTableController<T> controller, int dataIndex, T data) {
-    return infoRowHeightBuilder.call(dataIndex, data) ?? infoRowHeight ?? 0;
+  double infoRowHeight(FlexibleTableController<T> controller, int dataIndex, T data) {
+    return infoRowHeightBuilder.call(dataIndex, data) ?? fixedInfoRowHeight ?? 0;
   }
 }