+++ lib/src/selectable/selectable_column.dart
@@ -1,7 +1,5 @@
 import 'package:flexible_scrollable_table_view/src/constraint/flexible_table_column_width.dart';
 import 'package:flexible_scrollable_table_view/src/flexible_column.dart';
-import 'package:flexible_scrollable_table_view/src/flexible_table_configurations.dart';
-import 'package:flexible_scrollable_table_view/src/flexible_table_controller.dart';
 import 'package:flutter/widgets.dart';
 
 ///定制化的可选中的 Column
@@ -18,63 +16,32 @@ abstract class AbsSelectableColumn<T> extends AbsFlexibleColumn<T> {
   AbsFlexibleTableColumnWidth get columnWidth => selectableWidth;
 
   @override
-  Widget buildHeader(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-  ) =>
-      buildSelectableHeader(controller, configurations, parentWidth);
+  Widget buildHeader(BuildArguments<T> arguments) => buildSelectableHeader(arguments);
 
   ///构建可编辑时的表头
-  Widget buildSelectableHeader(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-  );
+  Widget buildSelectableHeader(BuildArguments<T> arguments);
 
   ///构建不可编辑时的表头
-  Widget buildUnSelectableHeader(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-  ) =>
-      SizedBox(width: unSelectableWidth.getColumnWidth(parentWidth));
+  Widget buildUnSelectableHeader(BuildArguments<T> arguments) =>
+      SizedBox(width: unSelectableWidth.getColumnWidth(arguments.parentWidth));
 
   @override
-  Widget buildInfo(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-    int dataIndex,
-    T data,
-  ) =>
-      buildSelectableInfo(controller, configurations, parentWidth, dataIndex, data);
+  Widget buildInfo(BuildArguments<T> arguments, int dataIndex, T data) =>
+      buildSelectableInfo(arguments, dataIndex, data);
 
   ///构建可编辑时的表信息
-  Widget buildSelectableInfo(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-    int dataIndex,
-    T data,
-  );
+  Widget buildSelectableInfo(BuildArguments<T> arguments, int dataIndex, T data);
 
   ///构建不可编辑时的表信息
-  Widget buildUnSelectableInfo(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-    int dataIndex,
-    T data,
-  ) =>
-      SizedBox(width: unSelectableWidth.getColumnWidth(parentWidth));
+  Widget buildUnSelectableInfo(BuildArguments<T> arguments, int dataIndex, T data) =>
+      SizedBox(width: unSelectableWidth.getColumnWidth(arguments.parentWidth));
 }
 
 class SelectableColumn<T> extends AbsSelectableColumn<T> {
   const SelectableColumn(
     super.id, {
     required this.selectableWidth,
-    this.unSelectableWidth = const FixedTableColumnWidth(0),
+    this.unSelectableWidth = const FixedWidth(0),
     required this.selectableHeader,
     this.unSelectableHeader,
     required this.selectableInfo,
@@ -93,41 +60,19 @@ class SelectableColumn<T> extends AbsSelectableColumn<T> {
   final Widget Function(int dataIndex, T data)? unSelectableInfo;
 
   @override
-  Widget buildSelectableHeader(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-  ) =>
-      selectableHeader;
+  Widget buildSelectableHeader(BuildArguments<T> arguments) => selectableHeader;
 
   @override
-  Widget buildUnSelectableHeader(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-  ) {
-    return unSelectableHeader ?? super.buildUnSelectableHeader(controller, configurations, parentWidth);
+  Widget buildUnSelectableHeader(BuildArguments<T> arguments) {
+    return unSelectableHeader ?? super.buildUnSelectableHeader(arguments);
   }
 
   @override
-  Widget buildSelectableInfo(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-    int dataIndex,
-    T data,
-  ) =>
+  Widget buildSelectableInfo(BuildArguments<T> arguments, int dataIndex, T data) =>
       selectableInfo.call(dataIndex, data);
 
   @override
-  Widget buildUnSelectableInfo(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-    int dataIndex,
-    T data,
-  ) {
-    return unSelectableInfo?.call(dataIndex, data) ??
-        super.buildUnSelectableInfo(controller, configurations, parentWidth, dataIndex, data);
+  Widget buildUnSelectableInfo(BuildArguments<T> arguments, int dataIndex, T data) {
+    return unSelectableInfo?.call(dataIndex, data) ?? super.buildUnSelectableInfo(arguments, dataIndex, data);
   }
 }