+++ lib/src/sortable/sortable_column.dart
@@ -1,7 +1,5 @@
 import 'package:flexible_scrollable_table_view/src/constraint/flexible_table_column_width.dart';
 import 'package:flexible_scrollable_table_view/src/flexible_column.dart';
-import 'package:flexible_scrollable_table_view/src/flexible_table_configurations.dart';
-import 'package:flexible_scrollable_table_view/src/flexible_table_controller.dart';
 import 'package:flutter/widgets.dart';
 
 ///可排序 Column
@@ -30,20 +28,8 @@ class SortableColumn<T> extends AbsSortableColumn<T> {
   final Widget Function(int dataIndex, T data) info;
 
   @override
-  Widget buildHeader(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-  ) =>
-      header;
+  Widget buildHeader(BuildArguments<T> arguments) => header;
 
   @override
-  Widget buildInfo(
-    FlexibleTableController<T> controller,
-    AbsFlexibleTableConfigurations<T> configurations,
-    double parentWidth,
-    int dataIndex,
-    T data,
-  ) =>
-      info.call(dataIndex, data);
+  Widget buildInfo(BuildArguments<T> arguments, int dataIndex, T data) => info.call(dataIndex, data);
 }